requirements.txt
langchain
langgraph
langchain_openai
python-dotenv
openai
sounddevice
soundfile